requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.6.5
backports.functools-lru-cache==1.5
billiard==3.3.0.23
celery==3.1.18
configparser==3.7.4
Django==1.11.21
django-background-tasks==1.2.0
django-cities==0.5.0.6
django-common-helpers==0.9.2
django-compat==1.0.15
django-countries==5.3.3
django-crispy-forms==1.7.2
django-cron==0.5.1
django-filter==1.1.0
django-model-utils==3.1.2
django-tables2==2.0.6
django-tables2-bootstrap4==2.0.0b5
enum34==1.1.6
futures==3.2.0
isort==4.3.16
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pkg-resources==0.0.0
pylint==1.9.4
pytz==2018.9
redis==2.10.3
schedule==0.6.0
singledispatch==3.4.0.3
six==1.12.0
swapper==1.1.0
tqdm==4.32.2
wrapt==1.11.1
